Added armDrive
armDrive.c is a curses based program to allow the user to control
the arm using the arrow keys. It is not yet finished. Currently
the arrow keys scroll the numbers. Need to add output to UART,
a more 'dynamic' speed control for the scrolling, and some
instructions on the screen.